Migrated DataTypeEditor to new Type Editing Infrastructure
The existing DataTypeEditor was migrated to an ITypeEditorPage and a new
shell DataTypeEditor is introduced. This is the basis for adding the
DescriptionEditor later.
