+++ plugins/org.eclipse.fordiac.ide.attributetypeeditor/src/org/eclipse/fordiac/ide/attributetypeeditor/editors/AttributeTypeEditor.java
@@ -527,9 +527,4 @@ public class AttributeTypeEditor extends EditorPart implements CommandStackEvent
 		setInputWithNotify(input);
 	}
 
-	@Override
-	public LibraryElement getEditedElement() {
-		return attributeDeclaration;
-	}
-
 }
