+++ plugins/org.eclipse.fordiac.ide.datatypeeditor/src/org/eclipse/fordiac/ide/datatypeeditor/editors/DataTypeEditorPage.java
@@ -1,0 +1,247 @@
+/*******************************************************************************
+ * Copyright (c) 2020 Johannes Kepler University, Linz
+ *               2021 Primetals Technologies Austria GmbH
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *   Daniel Lindhuber, Bianca Wiesmayr
+ *     - initial API and implementation and/or initial documentation
+ *   Muttenthaler Benjamin
+ *     - fixed reload of view if file on file system did change
+ *     - use new saveType method of AbstractTypeExporter
+ *     - replaced DataTypeListener by AdapterImpl
+ *     - keep a copy of the datatype object in the view, otherwise the content of the file is changed even the save button was not pressed
+ *   Lukas Wais
+ *     - enabled Save As
+ *******************************************************************************/
+
+package org.eclipse.fordiac.ide.datatypeeditor.editors;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.eclipse.core.resources.IMarker;
+import org.eclipse.core.runtime.IProgressMonitor;
+import org.eclipse.fordiac.ide.datatypeeditor.Messages;
+import org.eclipse.fordiac.ide.datatypeeditor.widgets.StructEditingComposite;
+import org.eclipse.fordiac.ide.gef.annotation.GraphicalAnnotationModel;
+import org.eclipse.fordiac.ide.model.data.AnyDerivedType;
+import org.eclipse.fordiac.ide.model.data.StructuredType;
+import org.eclipse.fordiac.ide.typeeditor.ITypeEditorPage;
+import org.eclipse.fordiac.ide.ui.imageprovider.FordiacImage;
+import org.eclipse.gef.commands.CommandStack;
+import org.eclipse.gef.commands.CommandStackEvent;
+import org.eclipse.gef.commands.CommandStackEventListener;
+import org.eclipse.gef.ui.actions.ActionRegistry;
+import org.eclipse.gef.ui.actions.RedoAction;
+import org.eclipse.gef.ui.actions.UndoAction;
+import org.eclipse.gef.ui.actions.UpdateAction;
+import org.eclipse.jface.action.IAction;
+import org.eclipse.jface.viewers.ISelection;
+import org.eclipse.swt.widgets.Composite;
+import org.eclipse.ui.IActionBars;
+import org.eclipse.ui.IEditorInput;
+import org.eclipse.ui.IEditorSite;
+import org.eclipse.ui.IWorkbenchPart;
+import org.eclipse.ui.PartInitException;
+import org.eclipse.ui.actions.ActionFactory;
+import org.eclipse.ui.part.EditorPart;
+import org.eclipse.ui.part.MultiPageEditorSite;
+
+public class DataTypeEditorPage extends EditorPart implements ITypeEditorPage, CommandStackEventListener {
+
+	private CommandStack commandStack;
+	private GraphicalAnnotationModel annotationModel;
+	private StructEditingComposite structComposite;
+
+	private ActionRegistry actionRegistry;
+	private final List<String> selectionActions = new ArrayList<>();
+	private final List<String> stackActions = new ArrayList<>();
+	private final List<String> propertyActions = new ArrayList<>();
+
+	@Override
+	public void stackChanged(final CommandStackEvent event) {
+		updateActions(stackActions);
+	}
+
+	@Override
+	public void dispose() {
+		getCommandStack().removeCommandStackEventListener(this);
+		getSite().getWorkbenchWindow().getSelectionService().removeSelectionListener(this);
+		getActionRegistry().dispose();
+
+		if (structComposite != null) {
+			structComposite.dispose();
+		}
+
+		super.dispose();
+	}
+
+	@Override
+	public void doSave(final IProgressMonitor monitor) {
+		// nothing to do for save
+	}
+
+	@Override
+	public void doSaveAs() {
+		// nothing to do for saveAs
+	}
+
+	@Override
+	public void init(final IEditorSite site, final IEditorInput input) throws PartInitException {
+		setInput(input);
+		setSite(site);
+		site.getWorkbenchWindow().getSelectionService().addSelectionListener(this);
+		getCommandStack().addCommandStackEventListener(this);
+		initializeActionRegistry();
+		setActionHandlers(site);
+		setPartName(Messages.DataTypeEditorPage_DataType);
+		setTitleImage(FordiacImage.ICON_DATA_TYPE.getImage());
+		if (getSite() instanceof final MultiPageEditorSite mpes) {
+			annotationModel = mpes.getMultiPageEditor().getAdapter(GraphicalAnnotationModel.class);
+		}
+	}
+
+	private void setActionHandlers(final IEditorSite site) {
+		final ActionRegistry registry = getActionRegistry();
+		final IActionBars bars = site.getActionBars();
+		String id = ActionFactory.UNDO.getId();
+		bars.setGlobalActionHandler(id, registry.getAction(id));
+		id = ActionFactory.REDO.getId();
+		bars.setGlobalActionHandler(id, registry.getAction(id));
+		id = ActionFactory.DELETE.getId();
+		bars.setGlobalActionHandler(id, registry.getAction(id));
+		bars.updateActionBars();
+	}
+
+	@Override
+	public boolean isDirty() {
+		return getCommandStack().isDirty();
+	}
+
+	@Override
+	public boolean isSaveAsAllowed() {
+		return true;
+	}
+
+	@Override
+	public void createPartControl(final Composite parent) {
+		if (getType() instanceof final StructuredType structType) {
+			structComposite = new StructEditingComposite(parent, commandStack, structType, annotationModel, getSite());
+			getSite().setSelectionProvider(structComposite.getSelectionProvider());
+		}
+	}
+
+	@Override
+	public void setFocus() {
+		structComposite.setFocus();
+	}
+
+	private CommandStack getCommandStack() {
+		return commandStack;
+	}
+
+	@Override
+	public void selectionChanged(final IWorkbenchPart part, final ISelection selection) {
+		if (this.equals(getSite().getPage().getActiveEditor())) {
+			updateActions(selectionActions);
+		}
+	}
+
+	private void createActions() {
+		final ActionRegistry registry = getActionRegistry();
+		IAction action;
+
+		action = new UndoAction(this);
+		registry.registerAction(action);
+		getStackActions().add(action.getId());
+
+		action = new RedoAction(this);
+		registry.registerAction(action);
+		getStackActions().add(action.getId());
+	}
+
+	@Override
+	public <T> T getAdapter(final Class<T> key) {
+		if (key == ActionRegistry.class) {
+			return key.cast(getActionRegistry());
+		}
+		// adapting to the command stack is needed for the undo/redo actions
+		if (key == CommandStack.class) {
+			return key.cast(getCommandStack());
+		}
+		return super.getAdapter(key);
+	}
+
+	private List<String> getStackActions() {
+		return stackActions;
+	}
+
+	@Override
+	public AnyDerivedType getType() {
+		return (AnyDerivedType) ITypeEditorPage.super.getType();
+	}
+
+	private void initializeActionRegistry() {
+		createActions();
+		updateActions(propertyActions);
+		updateActions(stackActions);
+	}
+
+	private void updateActions(final List<String> actionIds) {
+		final ActionRegistry registry = getActionRegistry();
+		actionIds.forEach(id -> {
+			final IAction action = registry.getAction(id);
+			if (action instanceof final UpdateAction updateAction) {
+				updateAction.update();
+			}
+		});
+	}
+
+	private ActionRegistry getActionRegistry() {
+		if (null == actionRegistry) {
+			actionRegistry = new ActionRegistry();
+		}
+		return actionRegistry;
+	}
+
+	@Override
+	public void reloadType() {
+		if (getType() instanceof final StructuredType structType) {
+			structComposite.setStructType(structType);
+		}
+	}
+
+	@Override
+	public void gotoMarker(final IMarker marker) {
+		// currently we don't support markers
+	}
+
+	@Override
+	public boolean outlineSelectionChanged(final Object selectedElement) {
+		// currently we don't have outline for data types so we don't need to do
+		// anything
+		return false;
+	}
+
+	@Override
+	public boolean isMarkerTarget(final IMarker marker) {
+		return false;
+	}
+
+	@Override
+	public void setCommonCommandStack(final CommandStack commandStack) {
+		this.commandStack = commandStack;
+	}
+
+	@Override
+	public Object getSelectableObject() {
+		return null;
+	}
+
+}
