+++ plugins/org.eclipse.fordiac.ide.datatypeeditor/src/org/eclipse/fordiac/ide/datatypeeditor/messages.properties
@@ -1,5 +1,5 @@
  ################################################################################
- # Copyright (c) 2020 Johannes Kepler University, Linz
+ # Copyright (c) 2020, 2024 Johannes Kepler University, Linz
  #
  # This program and the accompanying materials are made available under the
  # terms of the Eclipse Public License 2.0 which is available at
@@ -11,13 +11,4 @@
  #   Daniel Lindhuber
  #     - initial API and implementation and/or initial documentation
  ################################################################################
-MessageDialogTitle_SaveError=Saving not possible
-MessageDialogContent_SaveError=An error has occurred while saving
-ErrorCompositeMessage=Loading has failed
-MessageDialogTitle_OutsideWorkspaceError=Not in workspace
-MessageDialogContent_OutsideWorkspaceError=Files outside of workspace cannot be opened
-StructViewingComposite_Headline=STRUCT Editor
-StructAlteringButton_SaveAndUpdate=Save and Update
-StructAlteringButton_SaveAs=Save As
-StructAlteringDescription=The following blocks were affected by changing this struct:
-WarningDialog_StructNotSaved=The changes have not been saved!
+DataTypeEditorPage_DataType=Data type
