+++ plugins/org.eclipse.fordiac.ide.fbtypeeditor.doc/plugin.xml
@@ -1,12 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <?eclipse version="3.4"?>
 <plugin>
-   <extension
-         point="org.eclipse.fordiac.ide.typeeditor.typeEditorPage">
-      <editorPage
-            class="org.eclipse.fordiac.ide.fbtypeeditor.doc.editors.DescriptionEditor"
-            sortIndex="70">
-      </editorPage>
-   </extension>
-
+	<extension
+    		point="org.eclipse.fordiac.ide.typeeditor.typeEditorPage">
+		<editorPage
+        		class="org.eclipse.fordiac.ide.fbtypeeditor.doc.editors.DescriptionEditor"
+            	sortIndex="70">
+			<enablement>
+            	<instanceof
+                	value="org.eclipse.fordiac.ide.model.libraryElement.FBType">
+	            </instanceof>
+    	     </enablement>
+      	</editorPage>
+   	</extension>
 </plugin>
