+++ plugins/org.eclipse.fordiac.ide.systemmanagement.ui/src/org/eclipse/fordiac/ide/systemmanagement/ui/editors/AutomationSystemEditor.java
@@ -484,9 +484,4 @@ public class AutomationSystemEditor extends AbstractBreadCrumbEditor implements 
 		return selection;
 	}
 
-	@Override
-	public LibraryElement getEditedElement() {
-		return system;
-	}
-
 }
